Shared ancestor for RuleDefinitionDto and EffectiveRuleDetailsDto
