+++ client-api/src/main/java/org/sonarsource/sonarlint/core/clientapi/backend/rules/RuleDefinitionDto.java
@@ -20,7 +20,6 @@
 package org.sonarsource.sonarlint.core.clientapi.backend.rules;
 
 import java.util.Map;
-import java.util.Optional;
 import javax.annotation.Nullable;
 import org.sonarsource.sonarlint.core.commons.CleanCodeAttribute;
 import org.sonarsource.sonarlint.core.commons.ImpactSeverity;
@@ -29,52 +28,16 @@ import org.sonarsource.sonarlint.core.commons.Language;
 import org.sonarsource.sonarlint.core.commons.RuleType;
 import org.sonarsource.sonarlint.core.commons.SoftwareQuality;
 
-public class RuleDefinitionDto {
-  private final String key;
-  private final String name;
-  private final IssueSeverity defaultSeverity;
-  private final RuleType type;
-  private final CleanCodeAttribute cleanCodeAttribute;
-  private final Map<SoftwareQuality, ImpactSeverity> defaultImpacts;
+public class RuleDefinitionDto extends AbstractRuleDto {
   private final Map<String, RuleParamDefinitionDto> paramsByKey;
   private final boolean isActiveByDefault;
-  private final Language language;
 
-  public RuleDefinitionDto(String key, String name, IssueSeverity defaultSeverity, RuleType type, @Nullable CleanCodeAttribute cleanCodeAttribute,
-    Map<SoftwareQuality, ImpactSeverity> defaultImpacts, Map<String, RuleParamDefinitionDto> paramsByKey, boolean isActiveByDefault, Language language) {
-    this.key = key;
-    this.name = name;
-    this.defaultSeverity = defaultSeverity;
-    this.type = type;
-    this.cleanCodeAttribute = cleanCodeAttribute;
-    this.defaultImpacts = defaultImpacts;
+  public RuleDefinitionDto(String key, String name, IssueSeverity defaultSeverity, RuleType type,
+    @Nullable CleanCodeAttribute cleanCodeAttribute, Map<SoftwareQuality, ImpactSeverity> defaultImpacts,
+    Map<String, RuleParamDefinitionDto> paramsByKey, boolean isActiveByDefault, Language language) {
+    super(key, name, defaultSeverity, type, cleanCodeAttribute, defaultImpacts, language);
     this.paramsByKey = paramsByKey;
     this.isActiveByDefault = isActiveByDefault;
-    this.language = language;
-  }
-
-  public String getKey() {
-    return key;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public IssueSeverity getDefaultSeverity() {
-    return defaultSeverity;
-  }
-
-  public RuleType getType() {
-    return type;
-  }
-
-  public Optional<CleanCodeAttribute> getCleanCodeAttribute() {
-    return Optional.ofNullable(cleanCodeAttribute);
-  }
-
-  public Map<SoftwareQuality, ImpactSeverity> getDefaultImpacts() {
-    return defaultImpacts;
   }
 
   public Map<String, RuleParamDefinitionDto> getParamsByKey() {
@@ -83,9 +46,5 @@ public class RuleDefinitionDto {
 
   public boolean isActiveByDefault() {
     return isActiveByDefault;
-  }
-
-  public Language getLanguage() {
-    return language;
   }
 }
